Integer parsers moved to a separate file
